+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Add get_child_by_id and get_widget_by_id, remove get_child https://github.com/Textualize/textual/pull/1146
+- Add easing parameter to Widget.scroll_* methods https://github.com/Textualize/textual/pull/1144
 
 ### Changed